requirements.txt
websockets
python-dotenv==12.0
